+++ src/main/scala/org/broadinstitute/hail/driver/AnnotateVariants.scala
@@ -13,7 +13,6 @@ object AnnotateVariants extends SuperCommand {
   register(AnnotateVariantsBed)
   register(AnnotateVariantsExpr)
   register(AnnotateVariantsIntervals)
-  register(AnnotateVariantsJSON)
   register(AnnotateVariantsTable)
   register(AnnotateVariantsVCF)
   register(AnnotateVariantsVDS)