+++ src/main/scala/org/broadinstitute/hail/driver/Command.scala
@@ -48,7 +48,6 @@ object ToplevelCommands {
   register(AnnotateVariants)
   register(AnnotateGlobal)
   register(Cache)
-  register(ImportAnnotations)
   register(CompareVDS)
   register(Count)
   register(DownsampleVariants)
@@ -68,6 +67,7 @@ object ToplevelCommands {
   register(GRM)
   register(GQByDP)
   register(GQHist)
+  register(ImportAnnotations)
   register(ImportVCF)
   register(ImputeSex)
   register(LinearRegressionCommand)
@@ -142,7 +142,7 @@ abstract class SuperCommand extends Command {
 
   override def parseArgs(args: Array[String]): Options = {
     val options = newOptions
-    if (args(0) ==  "-h")
+    if (args(0) == "-h")
       options.printUsage = true
 
     val subArgs = args.dropWhile(_ == "-h")