+++ src/main/scala/org/broadinstitute/hail/variant/Variant.scala
@@ -7,7 +7,6 @@ import org.broadinstitute.hail.check.{Arbitrary, Gen}
 import org.broadinstitute.hail.expr._
 import org.json4s._
 
-import scala.collection.mutable
 import scala.math.Numeric.Implicits._
 
 object Contig {
@@ -120,6 +119,8 @@ case class AltAllele(ref: String,
     ("ref", JString(ref)),
     ("alt", JString(alt))
   )
+
+  override def toString: String = s"$ref/$alt"
 }
 
 object Variant {
@@ -205,8 +206,8 @@ case class Variant(contig: String,
   altAlleles: IndexedSeq[AltAllele]) extends Ordered[Variant] {
   /* The position is 1-based. Telomeres are indicated by using positions 0 or N+1, where N is the length of the
        corresponding chromosome or contig. See the VCF spec, v4.2, section 1.4.1. */
-  require(start >= 0, s"created a variant with negative position: `${this.toString}'")
-  require(!ref.isEmpty, s"created a variant with an empty ref string: `${this.toString}'")
+  require(start >= 0, s"invalid variant: negative position: `${this.toString}'")
+  require(!ref.isEmpty, s"invalid variant: empty contig: `${this.toString}'")
 
   def nAltAlleles: Int = altAlleles.length