+++ changes.md
@@ -1,4 +1,15 @@
-
+ - **Overhaul of `table` commands that parse text files.**
+    - optional type imputation from the first twenty lines.
+    - option `-s, --samplecol` in `annotatesamples table` has been replaced by `-e, --sample-expr`.  This no longer takes a column name, but instead takes an expr language transformation of the column names to produce the sample key.  This could still be a column name e.g. `-e Sample`, but it could also be something like `-e FID.split("_")[1]`. 
+    - option `-v, --vcolumns` in `annotatevariants table` has been replaced by `-e, --variant-expr`.  This no longer takes a column name, but instead takes an expr language transformation of the column names to produce the variant key.  This expression will use the new expr variant constructor, `variant(args)`.  As before, this can take an imputed `Variant` column, one string column `CHR:POS:REF:ALT` (which will look like `Variant(VariantColumn)`) or it can take four arguments of type `String, Int, String, (String || Array[String])`.  In the second case, this argument will look something like `Variant(Chr, Pos.toInt, Ref, Alt)`. 
+    - added a few more options to all `table` commands:
+    - `--comment`: optional argument.  Will skip any line starting in the given string.
+    - `--no-header`: indicates that the file(s) don't have a header, and columns will be read as "_0", "_1", ... "_N".  You can use the other arguments with these columns, e.g. `--sample-expr _2`.
+    - `--impute`: turns on type imputation
+  - `importannotations json` and `annotatevariants json` are gone, because you can replicate the functionality with `table` commands
+  - option `--code` added to all `table` commands, as well as `annotatevariants vds` and `annotatevariants vcf`.  This can be used as an alternative to `--root` to have more control over which annotations are selected and where they're going. 
+  - new command-level documentation for all these commands.  See the launch pages for [**`annotatevariants`**](docs/commands/AnnotateVariants.md),  [**`annotatesamples`**](docs/commands/AnnotateSamples.md), and  [**`annotateglobal`**](docs/commands/AnnotateGlobal.md)
+____
  
  - Added expr function `fet` to calculate p-values using Fisher's Exact Test.  Invoke this with `fet(count1, count2, count3, count4)` [see docs for details](docs/HailExpressionLanguage.md)